+++ ChangeLog.md
@@ -7,6 +7,8 @@
 - Allow commands to not have bindings but still be submitted to
   `Command.bindings`.
   ([#78](https://github.com/davep/textual-enhanced/pull/78))
+- Ensure that commands that have no default binding still appear in the help
+  screen. ([#78](https://github.com/davep/textual-enhanced/pull/78))
 
 ## v1.6.0